First implementation of (SQL) SELECT
Emulates the query `SELECT A, B FROM table WHERE C=1`

The table has 3 integers columns A B and C and is represented as a list
of tuples (its rows). The output is a list of tuples which first
coordinate is either 1 or 0 (1 iff the corresponding row is selected)
and which two other coordinates are the contents columns A and B for
this row.