First draft for the FilterSeq pattern
Current behaviour:

`FilterSeq(p) $ arr` returns a regular array with the same size as
`arr`. The first elements of this array are the elements of `arr` which
satisfy the predicate `p` **in the same order**, the end of the array is
not specified.

The next step will be to use unknown-length-arrays